requirements/local.txt
-r base.txt

django-debug-toolbar
django-extensions
pylint
pep8